+++ src/game/state_checkpoint/model_keyword.rs
@@ -25,6 +25,7 @@ pub(super) enum KeywordSnapshot {
     Shroud,
     Unleash,
     Intimidate,
+    Menace,
     Undying,
     Indestructible,
     AttacksEachCombatIfAble,